Move acceptance tests to acceptance module
