+++ community/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/NodeIndexSeekByRangeAcceptanceTest.scala
@@ -20,7 +20,7 @@
 package org.neo4j.internal.cypher.acceptance
 
 import org.neo4j.cypher.internal.compiler.v3_1.pipes.{IndexSeekByRange, UniqueIndexSeekByRange}
-import org.neo4j.cypher.{ExecutionEngineFunSuite, NewPlannerTestSupport}
+import org.neo4j.cypher.{ExecutionEngineFunSuite, NewPlannerTestSupport, SyntaxException}
 
 /**
   * These tests are testing the actual index implementation, thus they should all check the actual result.
@@ -929,7 +929,7 @@ class NodeIndexSeekByRangeAcceptanceTest extends ExecutionEngineFunSuite with Ne
     // Given
     val query = "MATCH (n:Label) WHERE n.prop >= [1, 2, 3] RETURN n.prop AS prop"
 
-    an[SyntaxException] should be thrownBy {
+    a [SyntaxException] should be thrownBy {
       executeWithAllPlannersAndCompatibilityMode(query).toList
     }
   }
